+++ src/core/Lua/Bridges/LuaMovementMetaDataBridge.cpp
@@ -20,9 +20,9 @@ extern "C"
         
 		Core::MovementState state = *(Core::MovementState*)luaL_checkudata( L, 1, UNIT_MOVEMENT_META_DATA_TYPE_META );
 
-		float speedToDesire = luaL_checknumber( L, 2 );
-		float acceleration = luaL_checknumber( L, 3 );
-		float deceleration = luaL_checknumber( L, 4 );
+		float speedToDesire = static_cast<float>(luaL_checknumber( L, 2 ));
+		float acceleration =  static_cast<float>(luaL_checknumber( L, 3 ));
+		float deceleration =  static_cast<float>(luaL_checknumber( L, 4 ));
 
 		Core::GameData::SetMovementDataForState( state, speedToDesire, acceleration, deceleration );
